requirements.txt
ipython
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
pytest
python-dotenv
ruff
scikit-learn
tqdm
typer
playwright
requests
selenium
python-dotenv
loguru
-e .
